Tighten NewAPI account test probe PR
Keep the PR focused by removing unrelated Playwright contract-test changes, and make the NewAPI chat probe fail closed when its SSE parser never sees a completion marker.
