+++ frontend/package.json
@@ -12,9 +12,7 @@
     "typecheck": "vue-tsc --noEmit",
     "test": "vitest",
     "test:run": "vitest run",
-    "test:coverage": "vitest run --coverage",
-    "test:e2e": "playwright test",
-    "test:e2e:contracts": "playwright test -g template"
+    "test:coverage": "vitest run --coverage"
   },
   "dependencies": {
     "@lobehub/icons": "^4.0.2",
@@ -37,7 +35,6 @@
     "xlsx": "^0.18.5"
   },
   "devDependencies": {
-    "@playwright/test": "1.51.0",
     "@types/dompurify": "^3.0.5",
     "@types/file-saver": "^2.0.7",
     "@types/mdx": "^2.0.13",
